# docs/api-ledger/openapi-snapshot.json
{
  "components" : {
    "schemas" : {
...
      "DeviceResource" : {
        "properties" : {
          "channel" : {
            "type" : "string"
          },
          "deviceId" : {
            "type" : "string"
          },
          "validFrom" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ProviderTankResource" : {
        "properties" : {
          "buyerCompanyId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "capacity" : {
            "format" : "double",
            "type" : "number"
          },
          "critical" : {
            "type" : "boolean"
          },
          "currentLevel" : {
            "format" : "double",
            "type" : "number"
          },
          "customerAccountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "devices" : {
            "items" : {
              "$ref" : "#/components/schemas/DeviceResource"
            },
            "type" : "array"
          },
          "fuelProductId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "levelObservedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "levelPercent" : {
            "format" : "double",
            "type" : "number"
          },
          "levelSource" : {
            "type" : "string"
          },
          "lowLevelPercent" : {
            "format" : "double",
            "type" : "number"
          },
          "name" : {
            "type" : "string"
          },
          "organizationId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "siteId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "siteName" : {
            "type" : "string"
          },
          "unit" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RegisterTankResource" : {
        "properties" : {
          "capacity" : {
            "format" : "double",
            "type" : "number"
          },
          "customerAccountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelType" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "initialLevel" : {
            "format" : "double",
            "type" : "number"
          },
          "name" : {
            "maxLength" : 150,
            "minLength" : 0,
            "type" : "string"
          },
          "siteId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "unit" : {
            "maxLength" : 20,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "capacity", "customerAccountId", "name" ],
        "type" : "object"
      },
...
      "TankResource" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "capacity" : {
            "format" : "double",
            "type" : "number"
          },
          "configurationVersion" : {
            "format" : "int32",
            "type" : "integer"
          },
          "currentLevel" : {
            "format" : "double",
            "type" : "number"
          },
          "customerAccountId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "fuelType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "levelSource" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "organizationId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "siteId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "unit" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateProviderTankResource" : {
        "properties" : {
          "autoGenerateEnabled" : {
            "type" : "boolean"
          },
          "channel" : {
            "maxLength" : 60,
            "minLength" : 1,
            "type" : "string"
          },
          "deviceId" : {
            "maxLength" : 120,
            "minLength" : 1,
            "type" : "string"
          },
          "fuelProductId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "lowLevelPercent" : {
            "format" : "double",
            "maximum" : 90,
            "minimum" : 0,
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Token JWT Bearer para autenticar solicitudes a la API",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "support@fuelguard.app",
      "name" : "Soporte de FuelGuard",
      "url" : "https://fuelguard.app/support"
    },
    "description" : "API de FuelGuard para gestionar abastecimiento, flota, entregas, inventario, telemetría IoT y pagos por organización.",
    "license" : {
      "name" : "Apache 2.0 (licencia)",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "FuelGuard Platform",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/provider/tanks/{tankId}" : {
      "get" : {
        "description" : "Mismo recurso que el listado. Solo tanques activos de compradores con vínculo explícito, pedidos o solicitudes del proveedor autenticado.",
        "operationId" : "getProviderTank",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankId",
          "required" : true,
          "schema" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProviderTankResource"
                }
              }
            },
            "description" : "Detalle del tanque"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Id inválido"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Sin proveedor autenticado"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Tanque inexistente, inactivo o ajeno"
          }
        },
        "summary" : "Consultar tanque vinculado",
        "tags" : [ "Tanques del distribuidor" ]
      },
      "put" : {
        "description" : "Solo compradores vinculados. Campos omitidos se conservan; deviceId exige channel. Reemplazar dispositivo cierra la asociación previa del canal y conserva su historia. No cambia propietario/capacidad.",
        "operationId" : "update_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankId",
          "required" : true,
          "schema" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProviderTankResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProviderTankResource"
                }
              }
            },
            "description" : "Tanque actualizado"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Datos inválidos"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Sin proveedor autenticado"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Tanque no vinculado o producto ajeno"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Dispositivo duplicado"
          }
        },
        "summary" : "Editar política, producto o dispositivo del tanque",
        "tags" : [ "Tanques del distribuidor" ]
      }
    },
...
    "/api/tanks" : {
      "get" : {
        "description" : "Devuelve los tanques asociados a la organización activa del usuario.",
        "operationId" : "listTanks",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TankResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Tanques devueltos."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TankResource"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          }
        },
        "summary" : "Listar tanques",
        "tags" : [ "Tanques" ]
      },
      "post" : {
        "description" : "Flujo secundario de contingencia: el flujo principal del distribuidor es POST /api/provider/tanks. Crea un tanque para una cuenta y un sitio de la organización activa. Puede hacerlo un usuario con membresía activa en la organización de la cuenta; en el flujo habitual, es el comprador (ROLE_BUYER) de una organización CUSTOMER. Un proveedor (ROLE_PROVIDER) de otra organización no puede usar una cuenta ajena. La cuenta y el sitio deben pertenecer a la organización activa; también se validan capacidad, unidad y nivel inicial.",
        "operationId" : "registerTank",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterTankResource"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Tanque creado."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El cuerpo es inválido, la cuenta o el sitio no pertenecen a la organización activa, o la cuenta, el sitio, la capacidad, la unidad o el nivel inicial incumplen las reglas del dominio."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "El equipo heredado indicado ya está asociado a un tanque."
          }
        },
        "summary" : "Registrar tanque",
        "tags" : [ "Tanques" ]
      }
    },
    "/api/tanks/{tankId}" : {
      "get" : {
        "description" : "Devuelve el tanque indicado únicamente cuando pertenece a la organización activa del usuario.",
        "operationId" : "getTank",
        "parameters" : [ {
          "in" : "path",
          "name" : "tankId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TankResource"
                }
              }
            },
            "description" : "Tanque devuelto."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TankResource"
                }
              }
            },
            "description" : "El usuario no está autenticado o no tiene una organización activa."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TankResource"
                }
              }
            },
            "description" : "El tanque no existe o pertenece a otra organización."
          }
        },
        "summary" : "Consultar tanque por identificador",
        "tags" : [ "Tanques" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Servidor de la API",
    "url" : "http://localhost:8080"
  } ],
...
}